requirements.txt
Django==4.2.11
django-ckeditor==6.7.1
pillow==10.2.0
